13-model_state_delete_a.py

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -1,20 +1,20 @@
 #!/usr/bin/python3
-# Deletes all State objects with a name containing
-# the letter a from the database hbtn_0e_6_usa.
-
+''' Deletes states in the database containing the letter a '''
 import sys
-from sqlalchemy import create_engine
+from model_state import Base, State
+from sqlalchemy import (create_engine)
 from sqlalchemy.orm import sessionmaker
-from model_state import State
 
-if __name__ == "__main__":
-    engine = create_engine("mysql+mysqldb://{}:{}@localhost/{}"
-                           .format(sys.argv[1], sys.argv[2], sys.argv[3]),
-                           pool_pre_ping=True)
+
+if __name__ == "__main__" and len(sys.argv) == 4:
+    engine = create_engine('mysql+mysqldb://{}:{}@localhost:3306/{}'
+                           .format(sys.argv[1], sys.argv[2], sys.argv[3]))
     Session = sessionmaker(bind=engine)
     session = Session()
 
-    for state in session.query(State):
-        if "a" in state.name:
-            session.delete(state)
+    to_delete = session.query(State).filter(State.name.like('%a%')).all()
+
+    for state in to_delete:
+        session.delete(state)
+
     session.commit()
